add more top-level functions
The theory with removing schemas is that you can just call
parse_file_with with a pair created already, and someone wanting to
extend a schema would just rename it. add_slice already doesn't add
something if it already exists, so writing a schema would just be not
adding any values to a key and making a decision after parsing the user
conifg.

Mostly because I'm a lazy fucker, but also because this is absolutely
impossible.