version bump: 1.0 🎉
basically just did a few things that needed to be done to bring this
silly little library to the current year. TryInto/TryFrom have been
stabilized and Error::description has been deprecated. no one uses
std::error::Error anyway so if you want to use this in your project just
wrap it in an enum variant or something.